Update the packaging of liblouis native dll and supporting files.
The native liblouis.dll file must be packaged in a path of the form
`runtimes/<RID>/native>` where `<RID>` is the runtime identifier of
the target platform the dll is built for - in this case win-x64. With
this done, NuGet will handle the dependency flow for projects that
reference the SharpLouis package.

The supporting tables files must also be flowed into the output of
referencing projects, so we use the default NuGet behavior of bundling
Content MSBuild Items into the `content/` directory of the package, as
well as telling NuGet that these pieces of content should be copied to
the output directory of the referencing project.

Details on these behaviors can be referenced at
https://learn.microsoft.com/en-us/nuget/reference/msbuild-targets.